Change storage class of Population used in Camouflage::run().
It had been a stack allocated local, but then other methods (particularly for the mouse handler) needed access to it. I had made a member pointer to it, which seemed backwards. Changed it to me a member shared_ptr which is now set to a heap allocated instance via make_shared in Camouflage::run().
Add functions: getPopulation(), setPopulation(), getStepAsString()