Add “t” command to save current tournament image.
Allows interactively saving image in same format as the automatic “every 19th step” saving.
Actual change in EvoCamoVsStaticFCD::waitForUserInput() but for the sake of EvoCamoVsLearningPredator which inherits from it.
Based on new PythonComms::tWasTyped().
Already considering adding a second command (or using “1” and “2”) to also allow saving the previous step tournament.